requirements.txt
python-telegram-bot==20.7
libtorrent==2.0.9
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
python-dotenv==1.0.0
aiofiles==23.2.1
tqdm==4.66.1
humanize==4.9.0
